VariationIndices implementation in device code.
It's time to finally check this in. In addition to a significant speed-up, I also
quashed some nasty bugs, some that also existed in the previous commit. So even
though compatible source counts are now different than before (little bit scary),
I have a bit more confidence that it's closer to being right than whatever I was
doing before, which I know was wrong (in a small number of cases).

I'd really like to get to the bottom of if/why compatibility source counts differ
from a non-device branch, like subsets.

Oh, and there is a ton of debug/diagnostic code in here now, a real mess. Leaving
it in because there may still be some work to do.