stalled attempt at moving ClueManager.ncResultMap to C++.
which pretty much required I move Validator.mergeNcListCombo to C++,
which was really slow (~600k native calls).

So I also moved the inner peco list of Validator.mergeNcListResults
to C++, which was still really slow (~800 calls, bigger wraps/unwraps).

So I could keep going and convert all of Validator to C++, but then
there's the issue of the "magic knownSrcMap result list". which, I'd
either have to wrap back to JS (slow again), or expose as a ObjectWrap
via C++. Which seems suspicious. Something like getNumKnownSrcMapEntries(),
getKnownSrcMapEntry().

Or, just convert most/all of the functions that actually use knownSrcMap
to C++. That'd include buildKnownNcSrcListMap, which is a win. And
i'm not sure how much of the other stuff matters. show-components uses
it, but I think there'd be a fast way to call it for that use case.

OH! one other cool thing here that I should probably split out if I
never decide to finish this: updated binding.gyp with a copy action
that copies the last-built plugin binary to the root of the build
directory, so I can just require ../../../build/experiment from JS
without having to worry about Debug/Release directories.