tentative fix for -t <primary-clue-name-variation>.
first change since returning from vacation, and things are a bit fuzzy.

but this was a difficult change to wrap my head around before I left, as well.

i should probably document and/or name and/or encapsulate some of the clue-manager
global data into classes to help me remember what is what and what goes where. its
a bit of a mess of rather poorly named global data and functions at the moment,
and i don't have a lot of intuition about it all, so could use the help of some
additional structure.