Discovered probable cause of result discrepancy between CPU and GPU when using three --or args.
I think it's related to not calling is_source_OR_compatible when the
smallest_spans XOR check returns All. Since that is exclusively an
XOR check, and says nothing about OR compatibility.

How exactly that results in the values I'm seeing I'm not sure. However,
looking more closely at the code that branches on the smallest_spans
result, I think I may need to introduce a shared variable and some
synchronizations to collate the correct compatibility result for all
threads in the block.

This is a checkpoint before I start doing that.