some preliminary work on a (potentially faster) version of the consistency checker.
still some work to be done to determine if a particular NC is VALID, which the
old version determines by running PreCompute on every source csv/name-list, and
which I could also do here on only a discovered subset of those, which would be
"faster" than the old version. And that might be the most direct way to do it.

I'm also wondering if this new logic isn't somehow applicable to solving the issue
 with -t not showing INVALID combos.