Mostly (ineffective) optimizations for, and comments about, validateSources.
It turns out that the real trick to this is just to reduce -m(ax_sources) to
below 20. Like, literally, 19. So that's what I did and now I probably don't
need to worry about this again for a while. When I do, there's some comments
to help me remember what the problem was, and a lot of unnecessarily complicated
code due to optimization attempts.