Refactored setPrimaryNameSrcIndicesMap() into init_primary_clues and several single-purpose functions.
This was the result of trying to figure out how to address the 'tricky issue' of disallowing multiple
of the same source names, and I saw some code that didn't make sense, and dealt with it, but made no
progress on my understanding of the tricky issue in the process.

I think I broke all src.ncLists in the previous commit's validateSources refactor. I need to either
see if I can fix that in the refactored code or revert to the old validator code.

-t two,words is broken as a result.