OR "kernel" maybe "working" now but without knowledge of (variation) compatible OR sources.
So it's very slow. Did a bunch of (probably unnecessary) work to process list-pairs and
compat-matrices in the merge kernel in a very specific order, such that "combo_indices"
could be generated & unpacked in filter kernel "intuitively", only to realize that there
is no such thing as a combo_idx. It's a complete fiction. There are only flat indices.

At least I have that knowledge and know how to "unpack" those now, so I guess that made
the effort worth it.

I need to create and pass in a struct for OR CompatIndices/Variations analagous to, but
different from, the SenteneceVariationIndices for XOR sources.

Working checkpoint.