OR source compat indices are now stored/allocated separately, not as part of variation_indices.
This is in preparation for using *actualy* variation indices for OR sources. In order to do that,
I had to templatize the VariationIndices structure by index type, since OR source variation
indices will not be fat indices. And that required I templatize a few other things as well like
IndexSpan, IndexSpanPair, FilterData::DeviceCommon.

Everything seems to still work.

fat indices.