Allocate and honor or_data.src_compat_results. Excellent perf increase in many cases.
However, there is a "non XOR-varition compatible" source case, that will still take
minutes for a single source to iterate over 100+ billion combinations of XOR and OR
sources, ultimately failing the compatibility check.

I think a solution to this is to generate src_compat_results for XOR sources as well,
which will help short-circuit the XOR-incompatible case.

It may also reduce the number of XOR-comparisons that take place overall, or at least
reach parity with VariationIndices, with a much smaller memory footprint.