Generate XOR src_compat_results, but not using them yet.
Spent a lot of time debugging a dumb bug.

Made a bunch of compatibility code generic for XOR/OR should I ever decide to use it for XOR.

Still tracking down the minutes-long 8090 test case failure, not sure how or if XOR
src_compat_results will help with that.

Good time for a checkpoint as everything seems to be working.