Fixed the incremental version of get_OR_compat_idx to be just as fast as binary search.
With 2 OR args. There's a "we're done" early exit condition that I'm not handling which
could potentially make it even faster. Also it's a 33% improvement on some of the nasty
no-compatible-OR-sources like listIndex 5497 (~36s -> ~24s). That may also be improved
with the early-exit.

Next up: integrate these changes into the 1-warp-producer implementaiton.