Create unique_variations for generated sources.
These variations are not in use yet for the purpose of source/xor variation
compat checking. But the src_idx indices passed to the kernel are (apparently)
all sorted correctly, and fair amount of work went in to the refactor to make
that happen.

2 and 3 OR-args pass memcheck. Big OR-arg run (no memmheck) fails with some
indexing error that I'll have to look into.

Partially-working checkpoint.