Fix previous indexing bug (caused by --stride 1): results array allocation size was wrong.
Remove -O3 from debug configuration in binding.gyp. This was preventing me from debugging
the indexing problem.

Added -Wconversion. Because it is probably the right thing to do, but my approach at
suppressing all of the warning was probably not the right way to do it.  Touched a lot of
files in the process, and kind of decided at the spur of the moment that "int" is better
for counts in the process. Unless it's device data. In some cases. In other words, my
guidelines are pretty flexible, which is probably not ideal.