Source <-> XOR unique varition indices are now generated but not used.
Just need to plug in some logic similar to the or-filter kernel to start using them,
but I want to dole the indices themselves out in "chunks", so that finished blocks
can pick up the work to assist long-running blocks.

It seems likely the DeviceUniqueVariations thing I did in FilterData probably isn't
necessary. But leaving it there just in case I'm wrong.