Some work resulting from investigating out-of-(device)-memory with -c2,15 and only --xor sources. Mostly related to better logging and debug messages.
The result of this investigation is that I think I need to do a better job of
detecting duplicate sources when they first occur (this probably means, when
they are "loaded"), in clue-manager.

An alternative approach would be to wait until loading is complete, then do a
2nd pass on the loaded clues and persist a (set of) "unique sources" data
structures, which might be cref's to the sources in the existing structures.

I haven't decided which approach to take, but I want to preserve the discovery
work I did up to this point, in case things get hairy and I need to revert.