Memory-utilization fixes. Revert compacting UV indices into separate results array.
Had some issues with unique sources implementation in KnownSources, so I parked that
work on another branch for now.

Merged in a few memory fixes, including elimination of repeated reallocation of
device_xor.src_compat_uv_indices every time filter_kernel was run, which was eating
memory.

Also turned off in-place compaction of UV indices, because it's obviously not right.
It's close though, but damn, that thing is hard to get right.